# Add highlight lexer cache to improve performance