@facelessuser 2023-07-22
We would have to look very close at this. I already see one issue. `language, lexer_options = self.get_extended_language(language)`. What you are proposing keeps a global cache, but if we have a script that spawns to different Markdown tasks, and both have different configs, the extended language could be different, but the cache will cause them to be the same. There could be other dynamic stuff, but I'm not sure.