@kamilkrzyskow 2023-07-22
I run the tox test now and it turned out that lexer guessing tests are failing 🤔, so perhaps I shouldn't save the cache when it guesses the language 😮‍💨. I shall investigate.