@kamilkrzyskow 2023-07-22
I missed the `language, lexer_options = self.get_extended_language(language)` part 👀 💦 
Between switching settings and profiling the speed it just slipped my mind that the language param of the function is later overwritten..., and I admit I probably don't understand what kind of edge-case there is when it comes to the extended language.

I moved things around to lookup cache after the `language` is overwritten, and cache based on the `language` used in the `get_lexer_by_name` function call, and I cache only after a successful function result.

Running tox with Python 3.11 locally results in a happy face 😄