@kamilkrzyskow 2023-07-22
I did some digging and it turned out that Pygments does indeed have some sort of cache in place https://github.com/pygments/pygments/blob/master/pygments/lexers/__init__.py#L126
And they store classes in the cache to call them later with different lexer options.
I don't know why it doesn't work properly on their end (as in adding the cache here adds performance), but I changed my approach to store the class and the performance gain seems to be somewhat the same.

I think that is my final edit ✌️