@kamilkrzyskow 2023-07-22
OK, I will look into how MkDocs processes the Markdown files and try to replicate the methodology of using the extensions. I will try to setup a script which simulates processing a thousand files with some generic C++ and Python code blocks. I believe this should be enough to showcase the performance difference, but let me know if you want something else.