@facelessuser 2023-07-22
We are testing a lot of things by going through MkDocs, you should be able to demonstrate an performance increase using Python Markdown directly.