@kamilkrzyskow 2023-07-22
So my initial hunch of the issue being related to my custom lexer turned out to be correct. Each time the `get_lexer_by_name` function is called it traverses the whole LEXERS dictionary and then if not found searches the plugin lexers... which means that in my case it always run the worst case scenario of traversing the whole 500+ lexer list and then finding the custom lexer at the end. Just my luck. https://github.com/pygments/pygments/blob/master/pygments/lexers/__init__.py#L122-L130

I made this script [performance_test.zip](https://github.com/facelessuser/pymdown-extensions/files/12137225/performance_test.zip) 
> It has markdown included, so the Github formatting breaks it

Running it on the e6474b38703b45e3dc431d3c1a0b1f24d80ee7fa commit results in 
```
Running bash_lexer x 100 times
-> Elapsed time 0.2318s
Running python_lexer x 100 times
-> Elapsed time 0.4003s
Running custom_lexer x 100 times
-> Elapsed time 2.5689s
```

and running it on the PR HEAD results in 
```
Running bash_lexer x 100 times
-> Elapsed time 0.2337s
Running python_lexer x 100 times
-> Elapsed time 0.3827s
Running custom_lexer x 100 times
-> Elapsed time 0.3471s
```

The `bash` lexer is ~40th in order and the `python` lexer is ~378th in order, therefore there is this difference in speed.
However, I don't understand why the added cache, doesn't seem to affect the Bash and Python times at all, it only affects the worst case scenario which is accessing custom plugin lexers (no matter if they exist or not) 🤔