@facelessuser 2023-07-22
Are you not registering your plugin as a proper package?