@kamilkrzyskow 2023-07-22
I'm not sure what do you mean by "registering as proper package". I copied the project structure of another custom lexer and it matches what is said in the documentation https://pygments.org/docs/plugins/
This is my setup.py with an appropriate entry point https://github.com/kamilkrzyskow/gothic-lexer/blob/main/setup.py#L17-L20
Pygments is able to find the lexer, however it seems that the lookup speed of custom plugin lexers isn't ideal.
I'll keep digging as there is an issue somewhere and maybe my proposed cache isn't the best way to solve it.