@facelessuser 2023-07-22
> This is my setup.py with an appropriate entry point https://github.com/kamilkrzyskow/gothic-lexer/blob/main/setup.py#L17-L20

Okay, just making sure. So custom lexers aren't cached and other lexers are? It was mentioned earlier that Pygments is supposed to be caching this. This makes me question if we are trying to patch this in the wrong place.